Making our first covariance matrix

+++ minerva/src/base/SystematicsBase.cpp
@@ -208,7 +208,7 @@ double * SystematicsBase::GetOptBinning(TTree * intree, const std::string &var_n
         int entries = intree->Draw(var_name.c_str(), sel.c_str() , "goff");
         // double best = start;
 
-        double delta = (double)(1. - (entries/dentry));
+        double delta = (double)(1. - ((double)entries/(double)dentry));
         cout << "entries/dentry = " << entries << "/" << dentry << " Starting delta = " << delta << endl;
         while( TMath::Abs(delta) < precision ){
         	start *= (1. + delta);
